devices/requirements.txt
# paho-mqtt
# loguru
# Django
# djangorestframework
# django-cors-headers
# paho-mqtt
# python-dotenv
# drf-spectacular
paho-mqtt
loguru
python-dotenv
getmac
numpy